added delete funciton for admin

+++ src/main/java/com/ncs/clinicmanagementsystem/restcontroller/UserRestController.java
@@ -24,81 +24,80 @@ import com.ncs.clinicmanagementsystem.service.UserService;
 public class UserRestController {
 
 	private UserService userService;
-	
 
 	@Autowired
 	public UserRestController(UserService theUserService) {
 		userService = theUserService;
 	}
-	
+
 	// expose "/user" and return list of users
 	@GetMapping("/user")
 	public List<User> findAll() {
-		
+
 		List<User> theUsers;
-		
+
 		try {
 			theUsers = userService.findAll();
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
 		}
-		catch(Exception e) {
-			throw new RuntimeException("Opps something happened. Please try again."); //throws error msg if error from db.
-		}
-		
-		if (theUsers.isEmpty()) { //throws error msg if users list is empty.
+
+		if (theUsers.isEmpty()) { // throws error msg if users list is empty.
 			throw new RuntimeException("Opps something happened. Please try again.");
 		}
-		
+
 		return theUsers;
 	}
-	
+
 	// add mapping for GET /user/{userEmail}
 	@GetMapping("/user/{userEmail}")
 	public User getUser(@PathVariable String userEmail) {
-		
+
 		User theUser;
-		
+
 		try {
 			theUser = userService.findByEmail(userEmail);
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
 		}
-		catch(Exception e) {
-			throw new RuntimeException("Opps something happened. Please try again."); //throws error msg if error from db.
+
+		if (theUser == null) { // throws error msg if user does not exist.
+			throw new RuntimeException("User email not found - " + userEmail);
 		}
-		
-		if (theUser == null) { //throws error msg if user does not exist.
-			throw new RuntimeException("User email not found - " + userEmail); 
-		}
-		
+
 		return theUser;
 	}
-	
+
 	// add mapping for POST /user - add new user
 	@PostMapping("/user/create")
 	public User addUser(@RequestBody User theUser) {
-		
+
 		// check if email already exist.
 		User tempUser;
 		try {
 			tempUser = userService.findByEmail(theUser.getEmail());
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
 		}
-		catch(Exception e) {
-			throw new RuntimeException("Opps something happened. Please try again."); //throws error msg if error from db.
-		}
-		
+
 		if (tempUser != null) { // Throw exception if email already exist.
 			throw new RuntimeException("This email address has already been registered.");
 		}
-		
+
 		// Salty password
 		theUser.setPassword(BCrypt.hashpw(theUser.getPassword(), BCrypt.gensalt()));
-			
+
 		// save new user to db.
 		try {
 			userService.save(theUser);
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
 		}
-		catch(Exception e) {
-			throw new RuntimeException("Opps something happened. Please try again."); //throws error msg if error from db.
-		}
-		
+
 		return theUser;
 	}
 
@@ -109,23 +108,24 @@ public class UserRestController {
 		String userPassword = theUser.getPassword();
 
 		User tempUser;
-		
+
 		try {
 			tempUser = userService.findByEmail(userEmail);
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
 		}
-		catch(Exception e) {
-			throw new RuntimeException("Opps something happened. Please try again."); //throws error msg if error from db.
-		}
-		
+
 		System.out.println(tempUser);
 
 		// If email does not exist in DB
 		if (tempUser == null) {
 			System.out.println("User not found");
-			
-			//throw new ResponseStatusException(HttpStatus.CONFLICT, "Email does not exist");
+
+			// throw new ResponseStatusException(HttpStatus.CONFLICT, "Email does not
+			// exist");
 			throw new RuntimeException("Email does not exist.");
-			//return null;
+			// return null;
 		}
 
 		boolean emailMatch;
@@ -144,23 +144,75 @@ public class UserRestController {
 			tempUser.setPassword("");
 			System.out.println("Wrong password");
 			throw new RuntimeException("Email or password does not match.");
-			//return null;
+			// return null;
 		}
 	}
 
-
-	// This mapping is temporary just to delete emails by giving the front-end an email address
+	// This mapping is temporary just to delete emails by giving the front-end an
+	// email address
 	// add mapping for DELETE /user/delete - delete a user
 	@DeleteMapping("user/delete")
 	public void delete(@RequestBody User theUser) {
+		User tempUser;
 		
 		try {
-			userService.deleteByEmail(theUser.getEmail());
+			tempUser = userService.findByEmail(theUser.getEmail());
+			System.out.println(tempUser);
+			if (tempUser == null) {
+				System.out.println("User not found");
+
+				// throw new ResponseStatusException(HttpStatus.CONFLICT, "Email does not
+				// exist");
+				throw new RuntimeException("Email does not exist.");
+				// return null;
+			} else {
+				try {
+					userService.deleteByEmail(theUser.getEmail());
+				} catch (Exception e) {
+					throw new RuntimeException("Opps something happened. Please try again.");
+				}
+			}
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
 		}
-		catch(Exception e) {
-			throw new RuntimeException("Opps something happened. Please try again.");
-		}
-		
+
 	}
-	
+
+	// This mapping is temporary just to delete emails by giving the front-end an
+	// email address
+	// add mapping for DELETE /user/delete - delete a user
+	@DeleteMapping("admin/delete")
+	public void deleteAsAdmin(@RequestBody User theUser) {
+		User tempUser;
+
+		try {
+			tempUser = userService.findByEmail(theUser.getEmail());
+			
+			if (tempUser == null) {
+				System.out.println("User not found");
+
+				// throw new ResponseStatusException(HttpStatus.CONFLICT, "Email does not
+				// exist");
+				throw new RuntimeException("Email does not exist.");
+				// return null;
+			} else {
+				try {
+					userService.deleteByEmail(theUser.getEmail());
+				} catch (Exception e) {
+					throw new RuntimeException("Opps something happened. Please try again.");
+				}
+			}
+		} catch (Exception e) {
+			throw new RuntimeException("Opps something happened. Please try again."); // throws error msg if error from
+																						// db.
+		}
+
+		System.out.println(tempUser);
+
+		// If email does not exist in DB
+
+
+	}
+
 }
